src/requirements.txt
Flask
transformers
torch
openai-whisper
librosa
numpy
nltk
werkzeug